+++ tests/_async/test_pool.py
@@ -1,5 +1,8 @@
+import typing as tp
+
 import httpcore
 import pytest
+import sniffio
 from httpcore._models import Request, Response
 
 import hishel
@@ -192,7 +195,7 @@ async def test_pool_with_cache_disabled_extension():
 
 
 @pytest.mark.anyio
-async def test_transport_with_custom_key_generator():
+async def test_pool_with_custom_key_generator():
     controller = hishel.Controller(key_generator=lambda request: request.url.host.decode())
 
     async with hishel.MockAsyncConnectionPool() as pool:
@@ -207,3 +210,23 @@ async def test_transport_with_custom_key_generator():
             response = await cache_transport.handle_async_request(request)
             assert response.extensions["from_cache"]
             assert response.extensions["cache_metadata"]["cache_key"] == "www.example.com"
+
+
+@pytest.mark.anyio
+async def test_pool_with_wrong_type_of_storage():
+    storage: tp.Union[hishel.FileStorage, hishel.AsyncFileStorage]
+
+    try:  # pragma: no cover
+        sniffio.current_async_library()
+        error = "Expected subclass of `Async" "BaseStorage` but got `FileStorage`"
+        storage = hishel.FileStorage()
+    except sniffio.AsyncLibraryNotFoundError:  # pragma: no cover
+        error = "Expected subclass of `BaseStorage` but got `Async" "FileStorage`"
+        storage = getattr(hishel, "Async" + "FileStorage")()
+
+    with pytest.raises(TypeError, match=error):
+        hishel.AsyncCacheConnectionPool(
+            pool=hishel.MockAsyncConnectionPool(),
+            controller=hishel.Controller(),
+            storage=storage,  # type: ignore
+        )